fix(Order) : price, totalPrice int 타입에서 BigDecimal 타입으로 변경

+++ src/main/java/com/sparta/deliverit/order/presentation/dto/response/CreateOrderInfo.java
@@ -3,6 +3,8 @@ package com.sparta.deliverit.order.presentation.dto.response;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 public class CreateOrderInfo {
     @JsonProperty("orderId")
     private final String orderId;
@@ -11,16 +13,16 @@ public class CreateOrderInfo {
     private final String orderStatus;
 
     @JsonProperty("totalPrice")
-    private final int totalPrice;
+    private final BigDecimal totalPrice;
 
     @Builder
-    private CreateOrderInfo(String orderStatus, String orderId, int totalPrice) {
+    private CreateOrderInfo(String orderStatus, String orderId, BigDecimal totalPrice) {
         this.orderStatus = orderStatus;
         this.orderId = orderId;
         this.totalPrice = totalPrice;
     }
 
-    public static CreateOrderInfo create(String orderStatus, String orderId, int totalPrice) {
+    public static CreateOrderInfo create(String orderStatus, String orderId, BigDecimal totalPrice) {
         return new CreateOrderInfo(orderStatus, orderId, totalPrice);
     }
 }
